baseline_v1_live/monitor_dashboard/requirements.txt
streamlit
pandas
numpy
pytz
plotly
psycopg2-binary
